Moved Arc, Node, and Graph classes to NGAT.Business.Domain.Core
